+++ dotnet/Personalizer/multislot-quickstart-v2PreviewSdk/README.md
@@ -6,7 +6,7 @@ This interactive sample takes the time of day and the users's device preference 
 
 ## Prerequisites
 
-The solution is a C# .NET Core console app project, so you will need [.NET Core 3.1](https://dotnet.microsoft.com/download/dotnet-core/3.1), and [Visual Studio 2019](https://visualstudio.microsoft.com/vs/) or [.NET Core CLI](https://docs.microsoft.com/en-us/dotnet/core/tools/).
+The solution is a C# .NET console app project, so you will need [.NET 5.0](https://dotnet.microsoft.com/download/dotnet/5.0), and [Visual Studio 2019](https://visualstudio.microsoft.com/vs/).
 
 ## Set up the sample
 
@@ -16,7 +16,7 @@ The solution is a C# .NET Core console app project, so you will need [.NET Core 
     git clone https://github.com/Azure-Samples/cognitive-services-quickstart-code.git
     ```
 
-- Navigate to _dotnet/Personalizer/multislot-quickstart_.
+- Navigate to _dotnet/Personalizer/multislot-quickstart-v2PreviewSdk_.
 
 
 ## Set up Azure Personalizer Service
