+++ dotnet/Personalizer/multislot-quickstart-v2PreviewSdk/multislot-quickstart.csproj
@@ -3,7 +3,10 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net5.0</TargetFramework>
-    <RootNamespace>multislot_quickstart</RootNamespace>
   </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Azure.AI.Personalizer" Version="2.0.0-beta.1" />
+  </ItemGroup>
 
 </Project>
