+++ dotnet/Personalizer/singleslot-quickstart-v1Sdk/Program.cs
@@ -1,0 +1,203 @@
+﻿// <Dependencies>
+using Microsoft.Azure.CognitiveServices.Personalizer;
+using Microsoft.Azure.CognitiveServices.Personalizer.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+// </Dependencies>
+
+namespace PersonalizerExampleSingleSlotSdkV1
+{
+    class Program
+    {
+        // <classVariables>
+        // The key specific to your personalizer resource instance; e.g. "0123456789abcdef0123456789ABCDEF"
+        private static readonly string ApiKey = "REPLACE-WITH-YOUR-PERSONALIZER-KEY";
+
+        // The endpoint specific to your personalizer resource instance; e.g. https://<your-resource-name>.cognitiveservices.azure.com
+        private static readonly string ServiceEndpoint = "https://REPLACE-WITH-YOUR-PERSONALIZER-RESOURCE-NAME.cognitiveservices.azure.com";
+        // </classVariables>
+
+        // <mainLoop>
+        static void Main(string[] args)
+        {
+            int iteration = 1;
+            bool runLoop = true;
+
+            // Get the actions list to choose from personalizer with their features.
+            IList<RankableAction> actions = GetActions();
+
+            // Initialize Personalizer client.
+            PersonalizerClient client = InitializePersonalizerClient(ServiceEndpoint);
+
+            do
+            {
+                Console.WriteLine("\nIteration: " + iteration++);
+
+                // <rank>
+                // Get context information from the user.
+                string timeOfDayFeature = GetUsersTimeOfDay();
+                string tasteFeature = GetUsersTastePreference();
+
+                // Create current context from user specified data.
+                IList<object> currentContext = new List<object>() {
+                    new { time = timeOfDayFeature },
+                    new { taste = tasteFeature }
+                };
+
+                // Exclude an action for personalizer ranking. This action will be held at its current position.
+                // This simulates a business rule to force the action "juice" to be ignored in the ranking.
+                // As juice is excluded, the return of the API will always be with a probability of 0.
+                IList<string> excludeActions = new List<string> { "juice" };
+
+                // Generate an ID to associate with the request.
+                string eventId = Guid.NewGuid().ToString();
+
+                // Rank the actions
+                var request = new RankRequest(actions, currentContext, excludeActions, eventId);
+                RankResponse response = client.Rank(request);
+                // </rank>
+
+                Console.WriteLine("\nPersonalizer service thinks you would like to have: " + response.RewardActionId + ". Is this correct? (y/n)");
+
+                // <reward>
+                float reward = 0.0f;
+                string answer = GetKey();
+
+                if (answer == "Y")
+                {
+                    reward = 1;
+                    Console.WriteLine("\nGreat! Enjoy your food.");
+                }
+                else if (answer == "N")
+                {
+                    reward = 0;
+                    Console.WriteLine("\nYou didn't like the recommended food choice.");
+                }
+                else
+                {
+                    Console.WriteLine("\nEntered choice is invalid. Service assumes that you didn't like the recommended food choice.");
+                }
+
+                Console.WriteLine("\nPersonalizer service ranked the actions with the probabilities as below:");
+                foreach (var rankedResponse in response.Ranking)
+                {
+                    Console.WriteLine(rankedResponse.Id + " " + rankedResponse.Probability);
+                }
+
+                // Send the reward for the action based on user response.
+                client.Reward(response.EventId, new RewardRequest(reward));
+                // </reward>
+
+                Console.WriteLine("\nPress q to break, any other key to continue:");
+                runLoop = !(GetKey() == "Q");
+
+            } while (runLoop);
+        }
+        // </mainLoop>
+
+        // <authorization>
+        /// <summary>
+        /// Initializes the personalizer client.
+        /// </summary>
+        /// <param name="url">Azure endpoint</param>
+        /// <returns>Personalizer client instance</returns>
+        static PersonalizerClient InitializePersonalizerClient(string url)
+        {
+            PersonalizerClient client = new PersonalizerClient(
+                new ApiKeyServiceClientCredentials(ApiKey)) { Endpoint = url };
+
+            return client;
+        }
+        // </authorization>
+
+        // <createUserFeatureTimeOfDay>
+        /// <summary>
+        /// Get users time of the day context.
+        /// </summary>
+        /// <returns>Time of day feature selected by the user.</returns>
+        static string GetUsersTimeOfDay()
+        {
+            string[] timeOfDayFeatures = new string[] { "morning", "afternoon", "evening", "night" };
+
+            Console.WriteLine("\nWhat time of day is it (enter number)? 1. morning 2. afternoon 3. evening 4. night");
+            if (!int.TryParse(GetKey(), out int timeIndex) || timeIndex < 1 || timeIndex > timeOfDayFeatures.Length)
+            {
+                Console.WriteLine("\nEntered value is invalid. Setting feature value to " + timeOfDayFeatures[0] + ".");
+                timeIndex = 1;
+            }
+
+            return timeOfDayFeatures[timeIndex - 1];
+        }
+        // </createUserFeatureTimeOfDay>
+
+        // <createUserFeatureTastePreference>
+        /// <summary>
+        /// Gets user food preference.
+        /// </summary>
+        /// <returns>Food taste feature selected by the user.</returns>
+        static string GetUsersTastePreference()
+        {
+            string[] tasteFeatures = new string[] { "salty", "sweet" };
+
+            Console.WriteLine("\nWhat type of food would you prefer (enter number)? 1. salty 2. sweet");
+            if (!int.TryParse(GetKey(), out int tasteIndex) || tasteIndex < 1 || tasteIndex > tasteFeatures.Length)
+            {
+                Console.WriteLine("\nEntered value is invalid. Setting feature value to " + tasteFeatures[0] + ".");
+                tasteIndex = 1;
+            }
+
+            return tasteFeatures[tasteIndex - 1];
+        }
+        // </createUserFeatureTastePreference>
+
+        // <createAction>
+        /// <summary>
+        /// Creates personalizer actions feature list.
+        /// </summary>
+        /// <returns>List of actions for personalizer.</returns>
+        static IList<RankableAction> GetActions()
+        {
+            IList<RankableAction> actions = new List<RankableAction>
+            {
+                new RankableAction
+                {
+                    Id = "pasta",
+                    Features =
+                    new List<object>() { new { taste = "salty", spiceLevel = "medium" }, new { nutritionLevel = 5, cuisine = "italian" } }
+                },
+
+                new RankableAction
+                {
+                    Id = "ice cream",
+                    Features =
+                    new List<object>() { new { taste = "sweet", spiceLevel = "none" }, new { nutritionalLevel = 2 } }
+                },
+
+                new RankableAction
+                {
+                    Id = "juice",
+                    Features =
+                    new List<object>() { new { taste = "sweet", spiceLevel = "none" }, new { nutritionLevel = 5 }, new { drink = true } }
+                },
+
+                new RankableAction
+                {
+                    Id = "salad",
+                    Features =
+                    new List<object>() { new { taste = "salty", spiceLevel = "low" }, new { nutritionLevel = 8 } }
+                }
+            };
+
+            return actions;
+        }
+        // </createAction>
+
+        // <readCommandLine>
+        private static string GetKey()
+        {
+            return Console.ReadKey().Key.ToString().Last().ToString().ToUpper();
+        }
+        // </readCommandLine>
+    }
+}
