+++ dotnet/Personalizer/singleslot-quickstart-v1Sdk/personalizer.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Azure.CognitiveServices.Personalizer" Version="1.0.0" />
+  </ItemGroup>
+
+</Project>
