+++ dotnet/Personalizer/singleslot-quickstart-v2PreviewSdk/personalizer.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Azure.AI.Personalizer" Version="2.0.0-beta.1" />
+  </ItemGroup>
+
+</Project>
